Set up simple functionable Option 4
Created a FileHandler Class instead of using General Class to handle reading lines and storing into a list from a file, need to discuss on this.

Also created a mergeSort function to sort the statements according to alphabetical order for Option 2.

Adjusted HashFunction in HashTable Class, need to relook at it.

Take note regarding the size of the HashTable. Should there be a preset high value or can it be adjustable.

Overall, do_4.py needs to be optimized.

Variable_expression function is repeated in both do_4.py and do_1.py, this is because the function of Option 1 is needed in Option 4.